requirements.txt
python-dotenv == 1.0.0
Flask == 3.0.0
pyodbc == 5.0.1
requests == 2.31.0